requirements.txt
pytube
moviepy
pydub
openai
python-dotenv